crontab

+++ rootfs/rootfs/usr/local/bin/green_create_zfs.sh
@@ -64,16 +64,16 @@ if [ -n "$MOUNT_DIR" ] && zfs_up && ( ! mountedOnGreen $ZFS_VOL ) ; then
      fi
    else
       zfs mount green/$ZFS_VOL
-      if [ $? != 0 ] ;then log_msg "zfs mount green/$ZFS_VOL ERROR" R ;else log_msg "zfs mount green/$ZFS_VOL OK" G ; fi 
+      if [ $? != 0 ] ;then log_msg "zfs mount green/$ZFS_VOL ERROR" R ;else log_msg "zfs mount green/$ZFS_VOL OK" G ; fi
    fi
 fi
 
 ZFS_VOL=opt_apps
 MOUNT_DIR=/opt/apps
 if [ -n "$MOUNT_DIR" ] && zfs_up && ( ! mountedOnGreen $ZFS_VOL ) ; then
+   rm -r -f $MOUNT_DIR
+   mkdir -p $MOUNT_DIR
    if ! volumeExistsOnGreen $ZFS_VOL ; then
-     rm -r -f $MOUNT_DIR
-     mkdir -p $MOUNT_DIR
      zfs create -o mountpoint=$MOUNT_DIR green/$ZFS_VOL
      if [ $? == 0 ] ; then
        log_msg "zfs  $POOL/$ZFS_VOL ; mounted on $MOUNT_DIR up :)" G
@@ -89,9 +89,9 @@ fi
 ZFS_VOL=home_docker
 MOUNT_DIR=${DOCKER_HOME_DIR}
 if [ -n "$MOUNT_DIR" ] && zfs_up && ( ! mountedOnGreen $ZFS_VOL ) ; then
+   rm -r -f $MOUNT_DIR
+   mkdir -p $MOUNT_DIR
    if ! volumeExistsOnGreen $ZFS_VOL ; then
-      rm -r -f $MOUNT_DIR
-      mkdir -p $MOUNT_DIR
       zfs create -o quota=$HOME_QUOTA -o mountpoint=$MOUNT_DIR green/$ZFS_VOL
       if [ $? == 0 ] ; then
          log_msg "zfs  $POOL/$ZFS_VOL ; mounted on $MOUNT_DIR up :)" G
@@ -106,10 +106,10 @@ fi
 
 ZFS_VOL=build
 MOUNT_DIR=/build
-if [ -n "$MOUNT_DIR" ] && zfs_up ( ! mountedOnGreen $ZFS_VOL ) ; then
+if [ -n "$MOUNT_DIR" ] && zfs_up && ( ! mountedOnGreen $ZFS_VOL ) ; then
+  rm -r -f $MOUNT_DIR
+  mkdir -p $MOUNT_DIR
   if ! volumeExistsOnGreen $ZFS_VOL ; then
-    rm -r -f $MOUNT_DIR
-    mkdir -p $MOUNT_DIR
     zfs create -o mountpoint=$MOUNT_DIR green/$ZFS_VOL
     if [ $? == 0 ] ; then
        log_msg "zfs  $POOL/$ZFS_VOL ; mounted on $MOUNT_DIR up :)" G
